requirements.txt
db-dtypes==1.3.0
docutils==0.21.2
functions-framework==3.*
google-api-core>=2.19.0
google-api-python-client>=2.137.0
google-auth>=2.34.0
google-cloud-bigquery>=3.23.0
google-cloud-bigquery-storage>=2.25.0
google-cloud-core>=2.4.1
google-cloud-logging>=3.11.2
google-cloud-storage>=2.3.0
more-itertools>=10.5.0
multiprocess>=0.70.17
numpy>=1.23.0
pandas>=1.3.0
python-dotenv>=1.0.0
requests>=2.26.0
vertexai>=1.70.0
